+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Binary/RoundTrip.hs
@@ -12,6 +12,10 @@ module Test.Cardano.Ledger.Conway.Binary.RoundTrip (
   roundTripConwayEraTypesSpec,
 ) where
 
+import Cardano.Ledger.Alonzo.Scripts (
+  AlonzoEraScript (..),
+  AsIx (..),
+ )
 import Cardano.Ledger.BaseTypes (StrictMaybe)
 import Cardano.Ledger.Binary (DecCBOR)
 import Cardano.Ledger.Compactible
@@ -32,6 +36,7 @@ roundTripConwayCommonSpec ::
   , EraGov era
   , EraStake era
   , EraCertState era
+  , AlonzoEraScript era
   , StashedAVVMAddresses era ~ ()
   , Arbitrary (Tx era)
   , Arbitrary (TxBody era)
@@ -43,6 +48,7 @@ roundTripConwayCommonSpec ::
   , Arbitrary (CompactForm (Value era))
   , Arbitrary (Script era)
   , Arbitrary (GovState era)
+  , Arbitrary (PlutusPurpose AsIx era)
   , Arbitrary (PParams era)
   , Arbitrary (PParamsUpdate era)
   , Arbitrary (PParamsHKD StrictMaybe era)